refactor(Scripts/ShatteredHalls): move Porung from SAI to cpp
